refactor #148: move round names to constants file
