+++ pages/api/scoreboard/index.ts
@@ -4,7 +4,7 @@ import { NextApiRequest, NextApiResponse } from 'next'
 import { formatRoundsData } from '@/utils/formatRoundsData'
 import { calculateTotals } from '@/utils/calculateTotals'
 import { findTopFive } from '@/utils/findTopFive'
-import { mockApiData } from '@/constants'
+import { mockApiData, roundNames } from '@/constants'
 import { calculateWinRate } from '@/utils/calculateWinRate'
 import {
   calculateCompletedChallenges,
@@ -99,8 +99,6 @@ export default async function handler(_: NextApiRequest, res: NextApiResponse) {
         playerStats.wins.total,
         playerStats.gamesPlayed.total,
       )
-
-      const roundNames = ['One', 'Two', 'Three', 'Four', 'Five'] as const
 
       roundNames.forEach((round) => {
         playerStats[`round${round}`] = formatRoundsData(RoundTimes, round)
