+++ pages/api/times/index.ts
@@ -1,7 +1,7 @@
 import { ApiPlayerStats, PlayerStats } from '@/interfaces/player'
 import { NextApiRequest, NextApiResponse } from 'next'
 import { formatRoundsData } from '@/utils/formatRoundsData'
-import { mockApiData } from '@/constants'
+import { mockApiData, roundNames } from '@/constants'
 import { removeBlacklistedPlayers } from '@/utils/removeBlacklistedPlayers'
 
 export default async function handler(
@@ -41,8 +41,6 @@ export default async function handler(
         name: PlayerName?.split('#')[0] || '',
         tag: PlayerName || '',
       }
-
-      const roundNames = ['One', 'Two', 'Three', 'Four', 'Five'] as const
 
       roundNames.forEach((round) => {
         playerStats[`round${round}`] = formatRoundsData(RoundTimes, round)
